LayoutTests/inspector/elements/user-properties.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="resources/user-properties.js"></script>
</head>

<body onload="runTest()">
<p>
Tests that DOMNode properly tracks own and descendants' user properties.
</p>

<div id="container">
  <div id="child1"></div>
  <div id="child2"><a href="#" id="aNode">Third-level node</a></div>
</div>

</body>
</html>
